Extract and move classes to domain layer
