+++ source/Energinet.DataHub.MeteringPoints.Domain/MasterDataHandling/MasterDataValue.cs
@@ -1,0 +1,58 @@
+﻿// Copyright 2020 Energinet DataHub A/S
+//
+// Licensed under the Apache License, Version 2.0 (the "License2");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
+using System;
+
+namespace Energinet.DataHub.MeteringPoints.Domain.MasterDataHandling
+{
+    public class MasterDataValue
+    {
+        public MasterDataValue(string name, Type valueType, Applicability applicability)
+        {
+            Name = name;
+            ValueType = valueType;
+            Applicability = applicability;
+        }
+
+        public string Name { get; }
+
+        public Type ValueType { get; }
+
+        public Applicability Applicability { get; private set; }
+
+        public object? Value { get; private set; }
+
+        public bool CanBeChanged { get; private set; } = true;
+
+        public void SetValue<T>(T value)
+        {
+            Value = value;
+        }
+
+        public void SetApplicability(Applicability applicability)
+        {
+            Applicability = applicability;
+        }
+
+        public void CanNotBeChanged()
+        {
+            CanBeChanged = false;
+        }
+
+        public bool HasRequiredValue()
+        {
+            return !(Applicability == Applicability.Required && Value is null);
+        }
+    }
+}
